Changed preferences import method: from module to Jitpack library - Updated smslibrary in order to have InvalidTelephoneNumberException - Moved RingtoneHandler in structure (as intended by group 4 project)
