+++ app/build.gradle
@@ -26,5 +26,6 @@ dependencies {
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
-    implementation project(path: ':preferences')
+    implementation 'com.github.CremaLuca:android-preferences:2.2'
+    compile project(path: ':smslibrary')
 }
