+++ settings.gradle
@@ -1,4 +1,3 @@
-include ':app', ':smslibrary', ':preferences'
+include ':app', ':smslibrary'
 rootProject.name='PhoneRinger'
 project(':smslibrary').projectDir = new File('library/smslibrary')
-project(':preferences').projectDir = new File('preferences')
